refactor: migrate logic to MVVM architecture
