+++ feature/src/main/java/com/example/feature/ui/onboarding/screen/CongratulationsScreen.kt
@@ -1,14 +1,12 @@
-package com.example.feature.ui.onboarding
+package com.example.feature.ui.onboarding.screen
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
