+++ feature/src/main/java/com/example/feature/ui/onboarding/viewmodel/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.feature.ui.onboarding
+package com.example.feature.ui.onboarding.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -6,12 +6,17 @@ import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.example.core.data.network.RetrofitInstance
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 class LoginViewModel : ViewModel() {
 
-    fun handleGoogleLogin(idToken: String?, navController: NavController) {
+    private val _loginSuccess = MutableStateFlow(false)
+    val loginSuccess: StateFlow<Boolean> = _loginSuccess
+
+    fun handleGoogleLogin(idToken: String?) {
         if (idToken == null) {
             Log.e("LOGIN", "idToken이 null입니다")
             return
@@ -19,7 +24,7 @@ class LoginViewModel : ViewModel() {
 
         Log.d("LOGIN", "idToken: $idToken")
 
-        viewModelScope.launch(Dispatchers.IO) {
+        viewModelScope.launch {
             try {
                 val response = RetrofitInstance.authService.getToken(idToken)
 
@@ -30,12 +35,7 @@ class LoginViewModel : ViewModel() {
                 if (!atk.isNullOrBlank()) {
                     RetrofitInstance.setAccessToken(atk)
                     Log.d("LOGIN", "ATK 설정 성공: $atk")
-
-                    withContext(Dispatchers.Main) {
-                        navController.navigate("홈") {
-                            popUpTo("Login") { inclusive = true }
-                        }
-                    }
+                    _loginSuccess.value = true
                 } else {
                     Log.e("LOGIN", "ATK가 비어 있거나 없음")
                 }
